いったんコミット

+++ trunk/shm-server/src/shm/controller/member/DeleteController.java
@@ -8,6 +8,9 @@ public class DeleteController extends MemberController {
     @Override
     protected Navigation runInTx() {
         
+        
+        assertAdminMember(); //管理者であるはず
+        
         memberDao.deleteObjectById(requestScope("key"));
         return null;
     }
